requirements.txt
appnope==0.1.0
backcall==0.1.0
beautifulsoup4==4.8.2
certifi==2018.8.24
decorator==4.4.1
Django==2.2
django-bootstrap4==1.1.1
importlib-metadata==1.3.0
ipdb==0.12.3
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.2
more-itertools==8.0.2
numpy==1.18.0
pandas==0.23.4
parso==0.5.2
path==13.1.0
path.py==12.4.0
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.5.2
python-dateutil==2.8.1
pytz==2019.3
scikit-learn==0.20.0
scipy==1.4.1
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
traitlets==4.3.3
wcwidth==0.1.7
zipp==0.6.0
